Add debug mode, refine bot user editing on season load.
The changing of a bot's username and avatar is heavily ratelimited. While testing, restarting the bot and changing seasons is required, and hitting these limits are typical. Instead, when in debug mode, the avatar isn't set and only the nickname is changed to prevent unnecessary account edit requests.

In the case that the rate limit is hit when not in debug mode, there's an added fallback to use the nickname instead.