Allows you to create Seasons. (#64)
* Allows you to create Season objects which change the bots behavior. For example, a season can determine things like the avatar, the nickname, and which cogs are loaded. Season automatically changes according to the date range you specify when you create it.

* removing some hungarian notation.

* Automatic season changes will now always happen at a minute past midnight, no matter when the bot started.

* catching dunders in the glob.

* Refine Season Creation behaviour and structure.

* Added channel and role constants, refactored roles into NamedTuples, added role check decorators from the main bot, and added role checks for the season change feature. Yes this is duplicate code from our main bot, but it will just have to be like that for now until we get a bot core running.

* replacing the or with an xor and switching out the assert for a UserWarning

* New lockfile

* changing discord.py to discord-py to prevent pip bug from putting two of them in our lockfile

* fixing flake errors

* flake8

* Cleaned everything up, but I seem to have introduced some sort of infinite load loop? o.O

* Fixing up all bugs in the halloween cogs. This should be ready for merge now.

* Add avatar_path baseclass method for consistency.

While making it simpler to add avatar urls in new season extensions, it also allows the avatar resource path to be changed in a single place if needed in future.

* Avoid shadowing builtin `object`.

* Add debug mode, refine bot user editing on season load.

The changing of a bot's username and avatar is heavily ratelimited. While testing, restarting the bot and changing seasons is required, and hitting these limits are typical. Instead, when in debug mode, the avatar isn't set and only the nickname is changed to prevent unnecessary account edit requests.

In the case that the rate limit is hit when not in debug mode, there's an added fallback to use the nickname instead.

* Add cancel load_seasons task on SeasonManager un/reload.

Previously the load_seasons task was loaded and looping forever. Even if the cog was unloaded for some reason, it would still be running. On loading the SeasonManager again, it would create a new load_seasons task, while the old one still existed.

Adding the cancellation allows the task to end when the cog is unloaded or reloaded, and will help assist with live code changes during development at a later time where it's possible to reload this extension (perhaps when the pending bot core is implemented).

* get_season_class helper, season class attribs, fix admin id

Changes `get_season`'s date check to not initialise unwanted classes (to avoid needless loading of tasks which would otherwise cause unexpected behaviour). To do this, defining attributes of season classes have been moved from `__init__` as an instance variable, to the class variable level. This also results in `__init__` not needing to be defined for the `SeasonBase` class, and `super().__init__()` not needing to be called in individual season classes, making things cleaner/simpler for them.

Adds a helper function for retrieving a season class and combines two unnecessarily separate if statements. Credits to @MarkKoz for the suggestions.

Reverts the admin ID mistakenly changed in a previous commit.

* Update bot/seasons/halloween/hacktoberstats.py

Co-Authored-By: heavysaturn <leon.haland@gmail.com>

* Update bot/seasons/halloween/halloween_facts.py

Co-Authored-By: heavysaturn <leon.haland@gmail.com>

* No more property in halloweenfacts

* Changed all aliases to tuples

* Made tokens a seperate namedtuple

* Update bot/seasons/halloween/spookyavatar.py

Co-Authored-By: heavysaturn <leon.haland@gmail.com>