Update LoveCalculator to address requested changes
This commit is addressing the new change requests on #100
Here's a list of what has been changed:

-> Moved `LOVE_DATA` and `LOVE_LEVELS` out of the LoveCalculator class:
This change has been made, so that the **love_matches.json** file does
not have to be loaded every time someone uses `.love`. Now the file is
loaded once when the cog is loaded.
Requested by: Scragly

-> Moved the Member converter to the function parameters:
I didn't know that this works. It's good to learn something new. This
also removes the `else` block, as it is no longer needed.
Requested by: Scragly

-> Removed the Roles instance:
Before these changes, an instance of `Roles` has been created inside the
if block: `ctx.guild.get_role(Roles().helpers).members`. This instance
has now been removed as it is unnecessary
Requested by: Scragly

-> Adjusted the embed's description's string length:
This project is aimed at 100 character lines in the future. The descrip-
tion is way longer, so its string has been split up to fit that specifi-
cation.
Requested by: Scragly