Fix LoveCalculator 0% bug
Recently a test of the .love command revealed an unintentional reaction
when 100% are reached. In previous versions this command had used a
gigantic conditional block, checking for the current percentage given by
the love_meter formula. When reaching 0 (meaning that the UserID's
evenly divided with 100) it would automatically convert that 0 to 100. A
big mistake from my side back then.

This has been fixed with these changes. By quickly modifying the formula
to use a mod 101 instead of a mod 100, we get an accurate percentage,
ranging from 0-100% instead of 0-99%. love_matches.json has also been
fixed to address these changes, meaning that the lowest entry now
actually is 0 and the highest 100.

TL;DR:
-> Edited love_meter formula to calculate mod 101 instead of mod 100
-> Edited lowest and highest entry in love_matches.json to be 0 and 100