Add .who_is_valentine and .valentine_facts
Closes #102

This pull request adds the commands who_is_valentine and valentine_facts.

The who_is_valentine command sends information about Saint Valentine of
Terni to the channel the command has been called in. This is done by making
use of the valentine_facts.json file's 'whois' attribute, which has
all the information stored in it.

The valentine_facts command sends a random fact with a random title to the
channel the command has been called in. This is done by making use of the
of the valentine_facts.json file's 'titles' and 'text' attribute, in which
both, the information for the titles and the facts are stored in.

Both commands do have docstrings, which explain what the command does in
an easy manner.

At the moment, there's a really sparse amount of facts available. I seek
to add a lot more in the near future.