Add drag queen names
The names are stored in a text file, since they're basically just a list, and
json seems like overkill for this case. In any case switching over to json
later would be easy. The names could be separated into first and last names in
order to generate more combinations, but that would break the wordplay that
many of the names have.

The names were taken from
https://www.fantasynamegenerators.com/drag-queen-names.php
as permitted by the footnote on the page.