Hotfix end date of Hacktober
The command is mostly used to calculate the time left in the DO's Hacktoberfest event we are participating in. However, the current end datetime was specified as a day before the actual end of the event. This hotfix remedies that for the end-calculation only.

This code should get a proper fix to rework the entire logic to use the correct start and end datetimes for DO's event.