Fix AOC countdown logic
The current time delta until the next AOC event assumes that the next event is next year's. While this is was a safe assumption when written, since the command would not be available until the season is loaded on December 1st, it provides an incorrect answer if the season is loaded prior. The logic has been adjusted to return the closest December 1st that is not in the past.

The feedback string has also been adjusted to give hours remaining if we're less than a day away from the event starting.