Fix Christmas season loading bug
The `SeasonBase.is_between_dates()` method assumes that the season's start and end dates are in the same year, which works for every season except for Christmas, since its end date is New Year's Day.

To fix this, the christmas.end() method has been overloaded to return the next year instead of the current year.

This does not affect checks made on 1 January of the next year, as the season's start date will guard against this.