Add correct emoji IDs to Evergreen reddit cog
The emoji IDs in the Reddit cog were for emojis that are not available to Seasonal Bot. While I think a proper solution using constants is to be preferred, I've edited in IDs for emojis that can be used by the bot in the mean time.

I've reused three of the emojis that we were already using for our Reddit webhook and added four new emojis to the `PyDis Emoji II` server. 

Reused emojis:
:user:              638729835442602003
:upvotes:           638729835245731840
:comments:          638729835073765387

New emojis available on PyDis Emoji II:
:reddit_logo:       676030265734332427
:reddit_post_photo: 676030265734201344
:reddit_post_text:  676030265910493204
:reddit_post_video: 676030265839190047

The latter four may need to be restyled to match the first three in the future.