Cycle icons within season in configured interval only
The `load_seasons` task now has an internal state, which increments for
every day. Before cycling the icon within the same season, first check
whether we've waited the configured amount of days since the last cycle.

Entering a new season, or changing the icon, resets the state to 0.

This allows us to slow down the rate at which we cycle icons, addressing
a bug where the icon wasn't loading for some users.