Deseasonify: fallback on evergreen on incomplete branding
The previous implementation would simply ignore assets not present in
the current season's directory.

On the condition that at least one asset is missing, and the current
season is not the evergreen, we also poll the evergreen directory,
and use it as fallback.

This behaviour is shielded such that we only make the extra API call if
there is a missing asset, i.e. if there is something to be gained.