Deseasonify: refer to bot icons as avatars
There is an annoying collision in the "server icon" vs "bot icon"
nomenclature, so I made the decision to always refer to the bot's image
as avatar. This choice is already visible through the module, we only
need to adjust the actual filename that we're looking for.

The filenames will be adjusted in the branding repository.