Deseasonify: remove excessive info embed information
Previously we were showing the entire queue of coming-up icons. As we're
now showing full paths (rather than just filenames), this creates a lot
of visual clutter that is unnecessary. It's not critical to know in
which order the icons will be cycled, as there's no way for us to affect
it anyway. This list of all available icons is still displayed above.

The information about configured cycle frequency goes in the footer,
if we're cycling.