Deseasonify: turn legacy docstrings into attrs
We no longer use the class docstrings for announcements, and they
mostly contain outdated information. Sentences still relevant are
used to populate the `description` attr (which shows in the branding
embed), the rest is scrapped.

The descriptions themselves can still be improved in the future, once
we figure out exactly in which direction we want to go with them.

The idea of providing a description for each season's branding was
originally brought up by neonsea, co-authored below.

Co-authored-by: Rasmus Moorats <me@neonsea.uk>