Deseasonify: improve branding embed
This improves both the visual embed, and the code used to generate it.

We'll now only display active months if the current season isn't the
evergreen - no reason to list the 12 months. The season's name and
avatar will be shown in the author field, as this is more less
aggressive and more visually pleasing than using title + thumbnail.
The embed will be coloured using the season's colour attr.

All values have safe fall-backs for situations where the seasonal
assets have failed to load.