Deseasonify: simplify `get_extensions` convenience function
This is a little tricky, as we only want to walk the modules in each
season's package. The pkgutil module provides a convenient function
called `walk_packages` which can walk recursively, but this isn't
desirable as not all modules are cogs - for example, the modules nested
under `evergreen.snakes` are only helpers, and the `setup` function is
instead present in the package's `__init__.py`.

As we no longer have modules directly in the seasons package, we can
remove the `ispkg` check.