Deseasonify: add daemon command group
This allows manually starting and stopping the daemon if necessary.
Enables introspection in the sense that we can check whether the daemon
is currently running, and for when is the next refresh scheduled.

In the case that the daemon dies due to an unhandled exception, the
start command is able to revive it. The daemon currently does not alert
us when it dies.