Deseasonify: do not automatically `apply` after `set`
The previous design was such that calling `set` with a new season will
automatically apply the branding. This meant 'previewing' a season
would require a separate command. Since `apply` is an expensive
operation with consequences, it appears logical to require
that the user invokes it separately & explicitly.

See docstring for further details.