Deseasonify: return failed assets
As the `apply` method attempts to apply all 4 types of assets, it is
useful to know which (if any) have failed. As all the methods that we
delegate to already return a bool denoting whether the operation has
succeeded, we can simply collect this information and return it to
the caller.