Deseasonify: add exceptions module
When loading extensions, d.py circumvents the import cache
by force-reloading the target module.

As a consequence, exception class declaration get re-executed,
which may cause the error handler to not recognize the exception via
its `isinstance` checks - the raised exception may not be an instance
of the class that the error handler originally imported.

For this reason, it is safer to declare the exceptions on the side,
in a module that will not be reloaded by d.py.