Deseasonify: abdicate responsibility to wait until bot is ready
The wrapper will no longer wait for the bot to be ready before it calls
the wrapped function for the first time. If the function requires the
bot's cache to be ready, it is responsible for awaiting the method
itself.

This removes the need to acquire a reference to the bot instance inside
the decorator, which seems to be difficult to do cleanly.

As Mark adds, this may in fact be safer as the bot may temporarily
disconnect while the task is active, and awaiting the bot's
ready status every time will prevent issues in such a situation.

Co-authored-by: MarkKoz <kozlovmark@gmail.com>