Deseasonify: show help on group invoke
Previously, when a group was called without a specific subcommand,
the status / info command was invoked automatically. This was
inconsistent with how other cogs behave, and makes it difficult
to discover the full API.

The logic from the `daemon` group is split off into a separate cmd
to allow this behaviour.