Deseasonify: add persistent branding configuration
A persistent config file will remember whether the daemon is supposed
to be running, or not. This means that instead of relying on the env
variable, the daemon can be turned on or off "once and for all"
via a command, and will not violate this decision on restart.

This effectively deprecates the `Branding.autostart` env var. It is
believed that having two ways to configure the same thing would
be more confusing than useful. The env var will be removed.

The two helper functions for reading and writing have been implemented
in a generic manner, meaning that should a new config key be added in
the future, the helpers shouldn't require any further adjustments.

Suggested by lemon and scragly.

Co-authored-by: Leon Sandøy <leon.haland@gmail.com>
Co-authored-by: scragly <29337040+scragly@users.noreply.github.com>