Add `wait_until_guild_available` coroutine
See docstring for explanation & comparison to the `on_ready` event.

This is taken from the Python bot repository. Originally implemented
by Mark. The `on_guild_available` callback was adapted and simplified
from the original implementation, as with Sentry in place, it is
believed that dispatching an error webhook is unnecessary.

Co-authored-by: MarkKoz <kozlovmark@gmail.com>