Remove SpookySound Cog that played sounds in voice
We had an old Cog that would allow our members to run a command to make
Seasonal Bot join a voice channel to play a spooky sound. However, as
our voice channel use has changed over the past year, we don't think
that it's still a good idea to do this.

That's why I removed the Cog and the constants related to it.