Migrate to GitHub Actions and GHCR
We are currently migrating our backend to a Kubernetes cluster. As this
requires changes to our CI anyway, this is a good time to also migrate
our CI Pipeline to GitHub Actions and start using the GitHub Container
Registry to push our builds.

Changes compared to the old situation:

- flake8 linting errors will now show up as GH Actions annotations
- containers are tagged with both "latest" and a short commit SHA

Note: This version of the workflow still pushes to DockerHub IN ADDITION
TO GHCR. This is to make the transition between the two deployment
systems as seamless as possible.