Fix lint errors introduced by persistence removal
We ripped some features out of the bot to make the migration to
kubernetes easier. However, that process left us with some linting
errors that will play up if anyone makes a PR.

I've added "noqa" to the valid TODO lines and commented out lines that
introduced now unused names or tried to use now undefined names. This
should have no influence on the bot, as it does not touch code that is
actually being run at the moment.