Split lint and build jobs into separate workflows
I've separated the lint and build jobs into two separate workflows:

- Lint:  .github/workflows/lint.yaml
- Build: .github/workflows/build.yaml

The main difference is that the Build workflow will be triggered if the
Lint workflow completes while it's been run on the "master" branch. The
build job will check if the Lint run was successful and if it were, it
actually builds the container, pushes it the GHCR, and triggers the
deployment to our kubernetes cluster.