Add constants parsing for multiple leaderboards
I've added a unified approach to setting data for multiple leaderboards
using environmental variables. Instead of setting separate variables for
the three pieces of data we have, hoping that the position of each board
matches up in the three fields, I now set one environmental variable,
AOC_LEADERBOARDS, that holds delimited data.

The data is in the format:

board_id1,session1,join_code1::board_id2,session2,join_code2[::...]

The staff leaderboards should be included as usual in this environmental
variable. Another environment variable, AOC_STAFF_LEADERBOARD_ID, can be
used to designate which leaderboard should be used as the staff board.

I've also made some other constants configurable in this commit and
added the role ID of the Events Lead role to allow the Events Lead to
force a reload of the leaderboard cache.