Add support for ignoring scores from specific days
I've added support for ignoring scores from specific days. A list of
days to ignore can be provided using the environment variable
`AOC_IGNORED_DAYS` as a comma-separated list.

This example would ignore day 1 and day 23:

AOC_IGNORED_DAYS=1,23

I've also added a helper function to sort the leaderboard not only on
the achieved score, but also on the number of stars an individual has
completed.