Let status update task sleep until the AoC starts
I've refactored the rich presence countdown task by making it hibernate
until 2 hours before the next Advent of Code starts if the task starts
up before the event has started. This ensures that the task will run
when the event starts and allows it to countdown to the first challenge.

After the event for the configured Advent of Code year has finished, the
task will terminate. This also means that the task will terminate
immediately in the year following the currently configured Advent of
Code; it will only start hibernating again once we configure the bot for
the next event.

No unnecessary, year-long hibernation.