Ensure status countdown waits for bot start-up
Trying to change the rich presence status of the bot too early in the
bot's start-up sequence will cause the task to fail. To make it wait,
I've added a `bot.wait_until_guild_available` point before the main loop
of the task starts.

This seems to solve the issue reliably, despite the `guild_available`
event not being directly related to when the bot's connection with the
API is ready to accept presence updates. However, the `ready` event is
know to fire too early, according to the discord.py community.