Update docstrings and fix grammar in comments
I've updated some docstrings to include more information about the inner
workings of some of the functions. In addition, I've also slightly
reformulated some block comments to improve their grammar.

Kaizen change: There was a redundant list comprehension in the Advent
of Code section of the constants. I've removed it.