Fix leaderboard glitch caused by duplicate names
We noticed that some entries on our leaderboard had an incorrect star
count attached to their name.

After a bit of digging, @HassanAbouelela discovered that this was caused
by the use of the member's name as the key for the leaderboard
dictionary: If different accounts used the same display name for the
leaderboard, they'd be combined into one glitched score dict.

The fix @HassanAbouelela wrote is to use the member id instead of the
name as the key for the leaderboard. I've changed a few names here and
there, but nothing major.

This commit closes #536