Disallow .aoc commands in primary aoc channel
Commands like `.aoc leaderboard` and `.aoc stats` proved to be spammy
in the main advent of code channel.
An aoc_commands channel has been added for aoc commands
and this update prohibits aoc commands from being used in the primary
aoc channel and adds the comands channel to the whitelist.

This also specifically allows the less spammier commands: join,
subscribe, unsubscribe, and countdown in the primary channel to foster
discussion though.