Add cog-level error handler for Incorrect Channel
If any of the "spammier" commands (stats, leaderboard) are used within
the primary advent of code channel, rather than a non-specific embed
we instead reply with the channel they should be using.

This also adds a "AOC_WHITELIST_PLUS" constant that makes it easier to
adjust what channels the non-spammier aoc commands can be used in.