Support a fallback session cookie in constants
To mitigate problems due to expiring session cookies, I'm currently in
the process of adding support for a fallback cookie. Basically, my
Advent of Code account is a member of *all* leaderboards, which means
that my cookie can be used to fetch all leaderboards as well.

As my session cookie should not expire until after the event, it should
not give us any issues. However, to avoid issues with issuing too many
requests from one session, we should still make sure to set individual
session values regardless of the mitigation.