Use workflow_run to send status embed to Discord
I've changed the way we send status embeds to make it work for PRs made
from forks without potentially exposing secrets. Instead of using the
initial workflows to send the embed, I've created a `workflow_run`
workflow that always runs in the context of the base repository.

And added benefit is that we don't have to add the status embed step to
two separate workflows.