Use Build Artifact to communicate PR information
A workflow run with a `workflow_run` payload does not contain the
necessary information to build a PR embed. As the old method of using
the API to fetch the relevant information turned out to be fragile (read
note 1 below) and the original Lint workflow already contains the
`pull_request` payload, we now store it as a build artifact.

The embed workflow then fetches the artifact and parses it to get the
relevant information out of it.

---

Note 1: Unfortunately, filtering Pull Requests using the "head"
parameter of the ``/repos/{owner}/{repo}/pulls` endpoint does not work
if the PR belongs to a fork with a different name than the base
repository: the API request will just return an empty array.

I've contacted GH to ask if this was intended or if it's a glitch, but,
for now, it's not a route that's easily available.