Add `MarkovPoemGenerator` cog
Fundamental changes:
- Bot can now process varying structures,
as opposed to just Shakespearean sonnets
- `self.bot.http_session` is now used,
instead of creating a new client session

Added custom exceptions:
- This allows the bot to not only stop in specific instances,
but also inform the developer why
- Added a `TimeoutError`in case the bot takes too long.
This is a fail-safe measure if the API timeout
and while-loop escape doesn't work.

Improving efficiency:
- Added `rhymeless_words` set so that the bot
does not have to perform an API request to find out again
- If a word is found to have no rhymes, it will fail on the spot
if is supposed to have another rhyming sentence.
This can be replaced to try to create a new line in the near future.
- A limiter has been added to the while-loop in `_get_rhyming_line`
- Timeouts have been (to the command and the API request)
in case of a bug

Miscallaneous changes:
- Path() is used instead of open() so that the bot will work on windows
- Instead of using lists to store the rhymes of a word,
a set is used instead
- Embed is now used to display poem for better visual aesthetics